// hr-frontend/src/lib/api/generated/model/listPositionsParams.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type ListPositionsParams = {
search?: string;
organizationId?: number;
orgUnitId?: number;
gradeId?: number;
active?: boolean;
open?: boolean;
occupied?: boolean;
/**
 * @minimum 0
 */
page?: number;
/**
 * @minimum 1
 * @maximum 100
 */
size?: number;
sortBy?: string;
direction?: string;
};